+++ packages/apollo-client/src/__tests__/ApolloClient.ts
@@ -547,9 +547,7 @@ describe('ApolloClient', () => {
         `,
       });
 
-      expect(
-        (client.queryManager.dataStore.getCache() as InMemoryCache).extract(),
-      ).toEqual({
+      expect((client.cache as InMemoryCache).extract()).toEqual({
         ROOT_QUERY: {
           a: 1,
         },
@@ -565,9 +563,7 @@ describe('ApolloClient', () => {
         `,
       });
 
-      expect(
-        (client.queryManager.dataStore.getCache() as InMemoryCache).extract(),
-      ).toEqual({
+      expect((client.cache as InMemoryCache).extract()).toEqual({
         ROOT_QUERY: {
           a: 1,
           b: 2,
@@ -586,9 +582,7 @@ describe('ApolloClient', () => {
         `,
       });
 
-      expect(
-        (client.queryManager.dataStore.getCache() as InMemoryCache).extract(),
-      ).toEqual({
+      expect((client.cache as InMemoryCache).extract()).toEqual({
         ROOT_QUERY: {
           a: 4,
           b: 5,
@@ -615,9 +609,7 @@ describe('ApolloClient', () => {
         `,
       });
 
-      expect(
-        (client.queryManager.dataStore.getCache() as InMemoryCache).extract(),
-      ).toEqual({
+      expect((client.cache as InMemoryCache).extract()).toEqual({
         ROOT_QUERY: {
           a: 1,
           d: {
@@ -646,9 +638,7 @@ describe('ApolloClient', () => {
         `,
       });
 
-      expect(
-        (client.queryManager.dataStore.getCache() as InMemoryCache).extract(),
-      ).toEqual({
+      expect((client.cache as InMemoryCache).extract()).toEqual({
         ROOT_QUERY: {
           a: 1,
           d: {
@@ -709,9 +699,7 @@ describe('ApolloClient', () => {
         `,
       });
 
-      expect(
-        (client.queryManager.dataStore.getCache() as InMemoryCache).extract(),
-      ).toEqual({
+      expect((client.cache as InMemoryCache).extract()).toEqual({
         ROOT_QUERY: {
           a: 1,
           b: 2,
@@ -765,9 +753,7 @@ describe('ApolloClient', () => {
         },
       });
 
-      expect(
-        (client.queryManager.dataStore.getCache() as InMemoryCache).extract(),
-      ).toEqual({
+      expect((client.cache as InMemoryCache).extract()).toEqual({
         ROOT_QUERY: {
           'field({"literal":true,"value":42})': 1,
           'field({"literal":false,"value":42})': 2,
@@ -810,9 +796,7 @@ describe('ApolloClient', () => {
         },
       });
 
-      expect(
-        (client.queryManager.dataStore.getCache() as InMemoryCache).extract(),
-      ).toEqual({
+      expect((client.cache as InMemoryCache).extract()).toEqual({
         ROOT_QUERY: {
           'field({"literal":true,"value":42})': 2,
           'field({"literal":false,"value":-1})': 1,
@@ -957,9 +941,7 @@ describe('ApolloClient', () => {
         `,
       });
 
-      expect(
-        (client.queryManager.dataStore.getCache() as InMemoryCache).extract(),
-      ).toEqual({
+      expect((client.cache as InMemoryCache).extract()).toEqual({
         foo: {
           __typename: 'Foo',
           e: 4,
@@ -995,9 +977,7 @@ describe('ApolloClient', () => {
         `,
       });
 
-      expect(
-        (client.queryManager.dataStore.getCache() as InMemoryCache).extract(),
-      ).toEqual({
+      expect((client.cache as InMemoryCache).extract()).toEqual({
         foo: {
           __typename: 'Foo',
           e: 4,
@@ -1027,9 +1007,7 @@ describe('ApolloClient', () => {
         `,
       });
 
-      expect(
-        (client.queryManager.dataStore.getCache() as InMemoryCache).extract(),
-      ).toEqual({
+      expect((client.cache as InMemoryCache).extract()).toEqual({
         foo: {
           __typename: 'Foo',
           e: 4,
@@ -1060,9 +1038,7 @@ describe('ApolloClient', () => {
         `,
       });
 
-      expect(
-        (client.queryManager.dataStore.getCache() as InMemoryCache).extract(),
-      ).toEqual({
+      expect((client.cache as InMemoryCache).extract()).toEqual({
         foo: {
           __typename: 'Foo',
           e: 4,
@@ -1112,9 +1088,7 @@ describe('ApolloClient', () => {
         fragmentName: 'fooFragment',
       });
 
-      expect(
-        (client.queryManager.dataStore.getCache() as InMemoryCache).extract(),
-      ).toEqual({
+      expect((client.cache as InMemoryCache).extract()).toEqual({
         foo: {
           __typename: 'Foo',
           e: 4,
@@ -1158,9 +1132,7 @@ describe('ApolloClient', () => {
         fragmentName: 'barFragment',
       });
 
-      expect(
-        (client.queryManager.dataStore.getCache() as InMemoryCache).extract(),
-      ).toEqual({
+      expect((client.cache as InMemoryCache).extract()).toEqual({
         foo: {
           __typename: 'Foo',
           e: 4,
@@ -1206,9 +1178,7 @@ describe('ApolloClient', () => {
         },
       });
 
-      expect(
-        (client.queryManager.dataStore.getCache() as InMemoryCache).extract(),
-      ).toEqual({
+      expect((client.cache as InMemoryCache).extract()).toEqual({
         foo: {
           __typename: 'Foo',
           'field({"literal":true,"value":42})': 1,
@@ -1391,9 +1361,7 @@ describe('ApolloClient', () => {
         bar: { __typename: 'Bar', d: 8, e: 9, f: 6 },
       });
 
-      expect(
-        (client.queryManager.dataStore.getCache() as InMemoryCache).extract(),
-      ).toEqual({
+      expect((client.cache as InMemoryCache).extract()).toEqual({
         foo: {
           __typename: 'Foo',
           a: 7,
@@ -1479,9 +1447,7 @@ describe('ApolloClient', () => {
         },
       });
 
-      expect(
-        (client.queryManager.dataStore.getCache() as InMemoryCache).extract(),
-      ).toEqual({
+      expect((client.cache as InMemoryCache).extract()).toEqual({
         ROOT_QUERY: {
           a: 1,
           b: 2,
@@ -1564,9 +1530,7 @@ describe('ApolloClient', () => {
         },
       });
 
-      expect(
-        (client.queryManager.dataStore.getCache() as InMemoryCache).extract(),
-      ).toEqual({
+      expect((client.cache as InMemoryCache).extract()).toEqual({
         ROOT_QUERY: {
           a: 1,
           b: 2,
@@ -1674,9 +1638,7 @@ describe('ApolloClient', () => {
         },
       });
 
-      expect(
-        (client.queryManager.dataStore.getCache() as InMemoryCache).extract(),
-      ).toEqual({
+      expect((client.cache as InMemoryCache).extract()).toEqual({
         ROOT_QUERY: {
           a: 1,
           b: 2,
@@ -1760,9 +1722,7 @@ describe('ApolloClient', () => {
         },
       });
 
-      expect(
-        (client.queryManager.dataStore.getCache() as InMemoryCache).extract(),
-      ).toEqual({
+      expect((client.cache as InMemoryCache).extract()).toEqual({
         ROOT_QUERY: {
           a: 1,
           b: 2,
@@ -1825,9 +1785,7 @@ describe('ApolloClient', () => {
         },
       });
 
-      expect(
-        (client.queryManager.dataStore.getCache() as InMemoryCache).extract(),
-      ).toEqual({
+      expect((client.cache as InMemoryCache).extract()).toEqual({
         ROOT_QUERY: {
           a: 1,
           b: 2,
@@ -1887,9 +1845,7 @@ describe('ApolloClient', () => {
         },
       });
 
-      expect(
-        (client.queryManager.dataStore.getCache() as InMemoryCache).extract(),
-      ).toEqual({
+      expect((client.cache as InMemoryCache).extract()).toEqual({
         ROOT_QUERY: {
           a: 1,
           b: 2,
@@ -1947,9 +1903,7 @@ describe('ApolloClient', () => {
         },
       });
 
-      expect(
-        (client.queryManager.dataStore.getCache() as InMemoryCache).extract(),
-      ).toEqual({
+      expect((client.cache as InMemoryCache).extract()).toEqual({
         ROOT_QUERY: {
           a: 1,
           b: 2,
@@ -2034,9 +1988,7 @@ describe('ApolloClient', () => {
         },
       });
 
-      expect(
-        (client.queryManager.dataStore.getCache() as InMemoryCache).extract(),
-      ).toEqual({
+      expect((client.cache as InMemoryCache).extract()).toEqual({
         ROOT_QUERY: {
           a: 1,
           b: 2,
@@ -2146,9 +2098,7 @@ describe('ApolloClient', () => {
         },
       });
 
-      expect(
-        (client.queryManager.dataStore.getCache() as InMemoryCache).extract(),
-      ).toEqual({
+      expect((client.cache as InMemoryCache).extract()).toEqual({
         ROOT_QUERY: {
           a: 1,
           b: 2,