+++ packages/apollo-client/CHANGELOG.md
@@ -3,6 +3,7 @@
 ### vNext
 - Support conditional refetches for mutations
 - Ensure network errors don't cause unhandled rejections on cache-and-network policies
+- Added the cache to the client for easier SSR
 
 ### 2.0.0-beta.5
 - Fix argument for FragmentMatcher